Update for my unique pointer
